requirements.txt
typing
fastapi
pydantic
mysql.connector
uvicorn[standard]
python-dotenv